added field-group to wrap around the fields
