+++ crayon-app/src/config/bit-config.js
@@ -45,10 +45,7 @@ const config = {
         grid: null
     },
     // This is how the modal grid will be displayed. The root array is the rows and the child array sets the columns. Use property name's.
-    modalGrid: [
-        ['name'],
-        ['type']
-    ],
+    modalGrid: null,
     /***************************************************************
     This is where we define all of the properties for an item.
     -----------------------------------------------------------
@@ -103,9 +100,9 @@ const config = {
             ]
         },
         {
-            name: 'photo',
+            name: 'photos',
             label: 'Photos',
-            type: 'photo',
+            type: 'text',
             many: true
         }
     ]
